build-clang: Switch to my fork of LLVM/Clang
Flash Clang is a rebranded CopperheadOS/HardenedOS Clang. Only reason
the vendor string isn't set to one of those is because this is a newer
version of Clang.

Make sure the local-init sanitizer patch[1] gets applied for anyone
using this script. Additionally, backport r332816[2] to Clang 6.0 so we
can disable ocaml bindings in the next patch.

[1]: https://github.com/AndroidHardeningArchive/platform_external_clang/commit/776a0955ef6686d23a82d2e6a3cbd4a6a882c31c
[2]: https://reviews.llvm.org/D42026

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>