env: Convert Distro ID logic into a function
To be picked up by elevated environment. opendoas support was recently added
to chkSudo(), which is an alternative to the widely-used sudo but doesn't
support preserving specific environment variable for root user.

This is to ensure that functions that need to read Distro ID will continue to
run on system that uses opendoas instead of sudo.

Signed-off-by: Albert I <kras@raphielgang.org>