archlinux: Update old-school method of updating CachyOS mirrorlist
It will now be ranked just once and copies be made over v3 and v4 mirrorlists
when any of them is available. This saves time since server lists are basically
the same.

R2 "mirror" is now left as is (at last, but for how long?), while REPO variable
is now renamed LEVEL since it's now being used solely for level-specific tasks.